server/ai-orchestrator-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
httpx>=0.25.0 # For making async HTTP requests to other services
# Add other dependencies later (e.g., redis, celery)
